:white_check_mark: Increase test coverage for log retrieval tests
